Transfer of function from global utils to local utils
